fully working software version

+++ TYPES.h
@@ -7,7 +7,7 @@
 
 typedef uint8_t BASE;			//data type for base (A, C, G or T)
 typedef uint16_t REF_INDEX;		//data type for index of the reference
-typedef uint8_t SEQ_INDEX;		//data type for index of the reference
+typedef uint16_t SEQ_INDEX;		//data type for index of the reference
 typedef uint32_t MATRIX_INDEX; 	//indexing the alignment matrix => ref_index size * seq_index size
 
 //SEQ and REF array types
